PycharmProjects/PythonProject/Secure File Transfer/requirements.txt
Flask==2.2.5
pycryptodome==3.17